ai_processing_azure_functions/requirements.txt
# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python
# azure-monitor-opentelemetry

azure-functions
openai
python-docx
pymupdf4llm
azure-ai-formrecognizer
requests  # For testing